backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
aiosqlite==0.20.0
pydantic[email]==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
python-dotenv==1.0.1
structlog==24.4.0
httpx==0.27.2
redis==5.2.0
minio==7.2.7
python-socketio==5.11.3
aiofiles==24.1.0
pytz==2024.2
pytest==8.3.3
pytest-asyncio==0.24.0
